# monotickets-api/docs/contracts/OPENAPI_monotickets.yaml
openapi: 3.0.0
paths:
...
  /events:
    post:
      operationId: EventsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventDto"
      responses:
        "201":
          description: Event created successfully
          content:
            application/json:
              schema:
                example:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  name: Boda de Juan y María
                  date: 2025-12-25T18:00:00.000Z
                  location: Jardín Las Rosas, CDMX
                  description: Celebración de nuestra boda
                  plannerId: 550e8400-e29b-41d4-a716-446655440001
                  active: true
                  createdAt: 2025-12-05T15:00:00.000Z
                  updatedAt: 2025-12-05T15:00:00.000Z
        "400":
          description: Invalid input
      security: &a3
        - bearer: []
      summary: Create a new event
      tags: &a4
        - events
    get:
      operationId: EventsController_findAll
      parameters:
        - name: plannerId
          required: false
          in: query
          description: Filter by planner ID
          schema:
            type: string
      responses:
        "200":
          description: Events retrieved successfully
      security: *a3
      summary: Get all events
      tags: *a4
  /events/{id}:
    get:
      operationId: EventsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event retrieved successfully
        "404":
          description: Event not found
      security: *a3
      summary: Get event by ID
      tags: *a4
    patch:
      operationId: EventsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventDto"
      responses:
        "200":
          description: Event updated successfully
        "404":
          description: Event not found
      security: *a3
      summary: Update event
      tags: *a4
    delete:
      operationId: EventsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event deleted successfully
        "404":
          description: Event not found
      security: *a3
      summary: Delete event (soft delete)
      tags: *a4
  /events/{id}/stats:
    get:
      operationId: EventsController_getStats
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event stats retrieved successfully
          content:
            application/json:
              schema:
                example:
                  totalGuests: 150
                  totalInvitations: 150
                  totalScans: 120
                  confirmedGuests: 140
      security: *a3
      summary: Get event statistics
      tags: *a4
  /events/{id}/publish:
    patch:
      operationId: EventsController_publish
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event published successfully
      security: *a3
      summary: Publish event
      tags: *a4
  /events/{id}/close:
    patch:
      operationId: EventsController_close
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event closed successfully
      security: *a3
      summary: Close event
      tags: *a4
  /events/{id}/block:
    patch:
      operationId: EventsController_block
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event blocked successfully
      security: *a3
      summary: Block event
      tags: *a4
  /events/{id}/unblock:
    patch:
      operationId: EventsController_unblock
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Event unblocked successfully
      security: *a3
      summary: Unblock event
      tags: *a4
...
  /director/dashboard:
    get:
      operationId: DirectorController_getDashboard
      parameters:
        - name: dateFrom
          required: false
          in: query
          description: Filter by date from
          schema:
            example: 2025-01-01T00:00:00.000Z
            type: string
        - name: dateTo
          required: false
          in: query
          description: Filter by date to
          schema:
            example: 2025-12-31T23:59:59.999Z
            type: string
        - name: plannerId
          required: false
          in: query
          description: Filter by planner ID
          schema:
            example: 550e8400-e29b-41d4-a716-446655440000
            type: string
        - name: role
          required: false
          in: query
          description: Filter by role
          schema:
            example: PLANNER
            type: string
            enum:
              - DIRECTOR_GLOBAL
              - PLANNER
              - STAFF
      responses:
        "200":
          description: Dashboard metrics retrieved successfully
          content:
            application/json:
              schema:
                example:
                  metrics:
                    totalEvents: 150
                    totalPlanners: 25
                    totalGuests: 5000
                    totalInvitations: 5000
                    totalScans: 4200
                    confirmedGuests: 4500
                    attendanceRate: 84
                    confirmationRate: 90
                  recentEvents: []
                  topPlanners: []
      security: &a12
        - bearer: []
      summary: Get global dashboard metrics
      tags: &a13
        - director
  /director/planners:
    get:
      operationId: DirectorController_getPlanners
      parameters: []
      responses:
        "200":
          description: Planners retrieved successfully
      security: *a12
      summary: Get all planners
      tags: *a13
  /director/planners/{id}:
    get:
      operationId: DirectorController_getPlannerDetail
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Planner detail retrieved successfully
        "404":
          description: Planner not found
      security: *a12
      summary: Get planner detail
      tags: *a13
...
  /director/events:
    get:
      operationId: DirectorController_getGlobalEvents
      parameters:
        - name: dateFrom
          required: false
          in: query
          description: Filter by date from
          schema:
            example: 2025-01-01T00:00:00.000Z
            type: string
        - name: dateTo
          required: false
          in: query
          description: Filter by date to
          schema:
            example: 2025-12-31T23:59:59.999Z
            type: string
        - name: plannerId
          required: false
          in: query
          description: Filter by planner ID
          schema:
            example: 550e8400-e29b-41d4-a716-446655440000
            type: string
        - name: role
          required: false
          in: query
          description: Filter by role
          schema:
            example: PLANNER
            type: string
            enum:
              - DIRECTOR_GLOBAL
              - PLANNER
              - STAFF
      responses:
        "200":
          description: Events retrieved successfully
      security: *a12
      summary: Get all events (global view)
      tags: *a13
  /director/stats:
    get:
      operationId: DirectorController_getStats
      parameters: []
      responses:
        "200":
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                example:
                  totalUsers: 30
                  totalEvents: 150
                  totalGuests: 5000
                  totalInvitations: 5000
                  totalScans: 4200
                  eventsByMonth: []
      security: *a12
      summary: Get global statistics
      tags: *a13
...
  /privacy/event/{eventId}:
    delete:
      operationId: PrivacyController_deleteEvent
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a20
...
info:
  title: Monotickets Platinum API
  description: API oficial para Monotickets Platinum (invited events, RSVPs,
    delivery y scanner).
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateEventDto:
      type: object
      properties:
        name:
          type: string
          description: Event name
          example: Boda de Juan y María
        date:
          type: string
          description: Event date
          example: 2025-12-25T18:00:00.000Z
        location:
          type: string
          description: Event location
          example: Jardín Las Rosas, CDMX
        description:
          type: string
          description: Event description
          example: Celebración de nuestra boda
        plannerId:
          type: string
          description: Planner user ID
          example: 550e8400-e29b-41d4-a716-446655440000
      required:
        - name
        - date
        - plannerId
    UpdateEventDto:
      type: object
      properties:
        name:
          type: string
          description: Event name
          example: Boda de Juan y María
        date:
          type: string
          description: Event date
          example: 2025-12-25T18:00:00.000Z
        location:
          type: string
          description: Event location
          example: Jardín Las Rosas, CDMX
        description:
          type: string
          description: Event description
          example: Celebración de nuestra boda
        plannerId:
          type: string
          description: Planner user ID
          example: 550e8400-e29b-41d4-a716-446655440000
...